Extract shared profile data onto its own page, unify page widths
Restructures around three pages, each with a single clear purpose:

- /profile — the data entered by every pathway: age, English,
  education, partner status, bonuses and nominated occupations. This
  is now the ONLY place that data is edited; it was previously
  duplicated across the points-tested page (as sections 01/02) and a
  read-only summary block on the sponsorship page.
- / (Independent Migration) — now just the calculated result, score
  timeline, reference info, fees and PR pathway, reading profile data
  read-only. The old 2-column sticky-sidebar layout is gone along
  with the section it was sized around; the result card now leads the
  page at a fixed compact width instead of stretching full-width.
- /sponsorship — drops its embedded "Your profile" block in favor of
  a one-line readout plus a link to /profile, matching how the other
  two pages now cross-reference it.

All three pages share the same flat max-width, fixing the visibly
different page widths between routes. The header nav (Profile /
Independent / Sponsorship) moved to its own row on every viewport —
inline with the logo it wrapped mid-word once a third tab was added
on narrow screens.

Also adds `persistDates()` (urlState.ts) so the Independent Migration
page — which now only ever edits `dates.visa491Grant` — can't clobber
a concurrent edit to shared/jobs made on the Profile page by writing
stale copies of data it no longer owns back to localStorage.
