+++ public/locales/en/common.json
@@ -5,8 +5,11 @@
   "chipHint": "View result",
   "placeholder": "Select",
   "navLabel": "Primary",
-  "navIndependent": "Independent Migration",
+  "navProfile": "Profile",
+  "navIndependent": "Independent",
+  "navSponsorship": "Sponsorship",
   "sections": {
+    "profile": "Profile",
     "shared": "Shared criteria",
     "jobs": "Skills assessments",
     "result": "Result",
@@ -16,6 +19,10 @@
     "fees": "Cost estimate",
     "sponsorship": "Employer sponsorship"
   },
+  "profilePageTitle": "Your Profile",
+  "profilePageSubtitle": "Shared across every pathway — age, English, education, partner status and nominated occupations",
+  "profileSeeResults": "See how this scores:",
+  "homeNoProfile": "Nothing entered yet.",
   "sharedNote": "These belong to you as the applicant and apply to every assessment: age, English, qualification, partner status, plus STEM, Australian study, regional study and NAATI bonuses.",
   "jobsNote": "An applicant may hold several skills assessments. Work experience and the Professional Year must relate to the nominated occupation, so they are entered per occupation.",
   "compareTitle": "COMPARE",
@@ -287,14 +294,10 @@
   "feesNominationFree": "free in at least one eligible state",
   "feesTotal": "Estimated total",
   "feesDisclaimer": "Figures are AUD estimates compiled 2026-07 from public fee schedules and may already be out of date — confirm current charges with the Department of Home Affairs, your assessing authority and the nominating state before budgeting.",
-  "spNote": "189/190/491 are points-tested; 482 and 186 are not — an employer sponsors you directly against a checklist (occupation gate, salary threshold, experience, age, English). This page checks the occupations, English level and age already entered on the Independent Migration page against all four streams — a different segment of the migration system worth knowing about if you can't yet reach 65 points.",
+  "spNote": "189/190/491 are points-tested; 482 and 186 are not — an employer sponsors you directly against a checklist (occupation gate, salary threshold, experience, age, English). This page checks the occupations, English level and age already entered on the Profile page against all four streams — a different segment of the migration system worth knowing about if you can't yet reach 65 points.",
   "spPageTitle": "Employer Sponsorship Checklist",
   "spPageSubtitle": "Visa subclasses 482 · 186 — a separate, non-points-tested pathway",
-  "spProfileTitle": "Your profile",
-  "spProfileEditLink": "Go to Independent Migration →",
-  "spProfileOccEmpty": "No occupation entered yet.",
-  "spProfileEnglishEmpty": "Not entered yet.",
-  "spProfileAgeEmpty": "Not entered yet.",
+  "spProfileEditLink": "Edit on Profile →",
   "spDetailsTitle": "Sponsorship details",
   "spEligibilityTitle": "Eligibility",
   "spHasSponsor": "I have (or a prospective employer has offered) sponsorship",
@@ -306,7 +309,7 @@
   "spTrtHint": "Temporary Residence Transition (TRT) pathway to subclass 186 — reduced from 3 to 2 years in Nov 2023; the occupation-list requirement was removed the same year.",
   "spAgeLine": "Age {{age}}",
   "spAgeBracketUnder45": "Age bracket confirms under 45",
-  "spAgeUnknown": "Age unknown — enter a birth month on the Independent Migration page to check the 186 age limit",
+  "spAgeUnknown": "Age unknown — enter a birth month on the Profile page to check the 186 age limit",
   "spEnglishOk": "English requirement met",
   "spEnglishUnknown": "English level not entered",
   "spEmpty": "No occupation entered yet — nothing to check yet.",
@@ -323,9 +326,9 @@
   "spGateSpecialistGroup": "Occupation is a Technician/Trades, Machinery Operator or Labourer role (excluded from Specialist Skills)",
   "spGateExperience": "Needs {{years}}+ years of relevant work experience",
   "spGateAgeLimit": "Must be under {{age}}",
-  "spGateEnglish": "Needs a recorded English level (Independent Migration page)",
+  "spGateEnglish": "Needs a recorded English level (Profile page)",
   "spGateTrtEligible": "Needs ≥2 years on a 482/457 with the nominating employer in this occupation",
-  "spDisclaimer": "A screening checklist only, not a nomination assessment — reuses the occupations, English level and age entered on the Independent Migration page (so CSOL-only occupations outside that list aren't checkable here), and approximates work experience from the Australian/overseas work brackets entered there rather than counting employment in the last 5 years exactly as the visa criteria define it. CSIT/SSIT and the 456-occupation CSOL are current as of 2026-07; some CSOL occupations carry extra salary or duty caveats not modelled here. Confirm eligibility with your employer and a registered migration agent before proceeding.",
+  "spDisclaimer": "A screening checklist only, not a nomination assessment — reuses the occupations, English level and age entered on the Profile page (so CSOL-only occupations outside that list aren't checkable here), and approximates work experience from the Australian/overseas work brackets entered there rather than counting employment in the last 5 years exactly as the visa criteria define it. CSIT/SSIT and the 456-occupation CSOL are current as of 2026-07; some CSOL occupations carry extra salary or duty caveats not modelled here. Confirm eligibility with your employer and a registered migration agent before proceeding.",
   "disclaimer": "This calculator is for reference only. Immigration rules change frequently — always verify with the Department of Home Affairs or a registered migration agent.",
   "cardFoot": "Reference only · verify with the Department of Home Affairs",
   "cardGoal": "Goal",
