+++ public/locales/zh/common.json
@@ -5,8 +5,11 @@
   "chipHint": "查看评估结果",
   "placeholder": "请选择",
   "navLabel": "主导航",
+  "navProfile": "个人资料",
   "navIndependent": "独立技术移民",
+  "navSponsorship": "雇主担保",
   "sections": {
+    "profile": "个人资料",
     "shared": "共同条件",
     "jobs": "职业评估",
     "result": "评估结果",
@@ -16,6 +19,10 @@
     "fees": "费用预估",
     "sponsorship": "雇主担保"
   },
+  "profilePageTitle": "个人资料",
+  "profilePageSubtitle": "各条移民路径通用——年龄、英语、学历、伴侣状态与提名职业",
+  "profileSeeResults": "查看评估结果：",
+  "homeNoProfile": "尚未填写任何信息。",
   "sharedNote": "以下条件属于申请人本人，与提名职业无关，所有职业评估共用：年龄、英语、学历、伴侣状态，以及 STEM、澳洲学习、偏远地区学习、NAATI 加分。",
   "jobsNote": "一位申请人可持有多个职业评估。工作经验与职业年须与提名职业相关，因此按职业分别填写，不可共用。",
   "compareTitle": "职业对比 · COMPARE",
@@ -287,14 +294,10 @@
   "feesNominationFree": "至少一个符合资格的州免费",
   "feesTotal": "预估总额",
   "feesDisclaimer": "以上金额为 2026 年 7 月根据公开费用标准整理的澳元估算，可能已过时——预算前请以澳大利亚内政部、评估机构及担保州的最新信息为准。",
-  "spNote": "189/190/491 采用打分制；482 和 186 不打分——由雇主直接担保，按清单逐项核对（职业清单门槛、薪资门槛、工作经验、年龄、英语）。本页用「独立技术移民」页面中已填写的职业、英语水平与年龄，对照全部四条担保路径——如果暂时凑不够 65 分，这是值得了解的另一条路径。",
+  "spNote": "189/190/491 采用打分制；482 和 186 不打分——由雇主直接担保，按清单逐项核对（职业清单门槛、薪资门槛、工作经验、年龄、英语）。本页用「个人资料」页面中已填写的职业、英语水平与年龄，对照全部四条担保路径——如果暂时凑不够 65 分，这是值得了解的另一条路径。",
   "spPageTitle": "雇主担保资格核对",
   "spPageSubtitle": "482 · 186 签证——独立于打分制之外的另一条路径",
-  "spProfileTitle": "个人资料",
-  "spProfileEditLink": "前往「独立技术移民」页面 →",
-  "spProfileOccEmpty": "尚未填写职业。",
-  "spProfileEnglishEmpty": "尚未填写。",
-  "spProfileAgeEmpty": "尚未填写。",
+  "spProfileEditLink": "前往「个人资料」页面修改 →",
   "spDetailsTitle": "担保信息",
   "spEligibilityTitle": "资格核对结果",
   "spHasSponsor": "已获得（或潜在雇主已提出）担保",
@@ -306,7 +309,7 @@
   "spTrtHint": "过渡到 186 永居的 TRT（Temporary Residence Transition）路径——工作年限已于 2023 年 11 月由 3 年降为 2 年，同年取消了职业清单门槛。",
   "spAgeLine": "年龄 {{age}}",
   "spAgeBracketUnder45": "年龄区间已确认在 45 岁以下",
-  "spAgeUnknown": "年龄未知——请在「独立技术移民」页面填写出生月份，以核对 186 的年龄上限",
+  "spAgeUnknown": "年龄未知——请在「个人资料」页面填写出生月份，以核对 186 的年龄上限",
   "spEnglishOk": "已满足英语要求",
   "spEnglishUnknown": "尚未填写英语水平",
   "spEmpty": "尚未填写职业——暂无可核对的内容。",
@@ -323,9 +326,9 @@
   "spGateSpecialistGroup": "该职业属于技术员/技工、机械操作员或劳工类（专业技能类别不适用）",
   "spGateExperience": "需要 {{years}} 年以上相关工作经验",
   "spGateAgeLimit": "须在 {{age}} 岁以下",
-  "spGateEnglish": "需要在「独立技术移民」页面填写英语水平",
+  "spGateEnglish": "需要在「个人资料」页面填写英语水平",
   "spGateTrtEligible": "需要持 482/457 在该职业为担保雇主工作满 2 年以上",
-  "spDisclaimer": "本页仅为资格初筛，并非提名评估——沿用「独立技术移民」页面中已填写的职业、英语水平与年龄（因此清单外的 CSOL 专属职业无法在此核对），并以该页填写的澳洲/海外工作经验区间估算工作年限，而非按签证要求精确计算「过去 5 年内」的从业经历。CSIT / SSIT 及 456 个 CSOL 职业均为 2026 年 7 月数据；部分 CSOL 职业还附带薪资或岗位限制条件，本工具未纳入建模。请在实际申请前与雇主及注册移民代理确认资格。",
+  "spDisclaimer": "本页仅为资格初筛，并非提名评估——沿用「个人资料」页面中已填写的职业、英语水平与年龄（因此清单外的 CSOL 专属职业无法在此核对），并以该页填写的澳洲/海外工作经验区间估算工作年限，而非按签证要求精确计算「过去 5 年内」的从业经历。CSIT / SSIT 及 456 个 CSOL 职业均为 2026 年 7 月数据；部分 CSOL 职业还附带薪资或岗位限制条件，本工具未纳入建模。请在实际申请前与雇主及注册移民代理确认资格。",
   "disclaimer": "本计算器仅供参考。移民政策经常变动，请以澳大利亚内政部官方网站或注册移民代理的信息为准。",
   "cardFoot": "仅供参考 · 请以澳大利亚内政部官方信息为准",
   "cardGoal": "目标",
