+++ src/app/HomeClient.tsx
@@ -1,13 +1,9 @@
 'use client';
 
-import { Suspense, memo, useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { Suspense, useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
+import Link from 'next/link';
 import Header from '@/components/Header';
-import SectionHeading from '@/components/SectionHeading';
-import SharedCriteriaSection from '@/components/SharedCriteriaSection';
-import JobCard from '@/components/JobCard';
-import type { JobUIState } from '@/components/JobCard';
-import ComparisonTable from '@/components/ComparisonTable';
 import ResultsBand from '@/components/ResultsBand';
 import ReferenceSection from '@/components/ReferenceSection';
 import Pr191Section from '@/components/Pr191Section';
@@ -21,12 +17,11 @@ import Footer from '@/components/Footer';
 import { useLocalStorage } from '@/hooks/useLocalStorage';
 import { useAnimatedNumber } from '@/hooks/useAnimatedNumber';
 import { evaluate } from '@/lib/points';
-import type { JobEvaluation } from '@/lib/points';
-import type { JobAssessment, PlanningDates, SharedCriteria } from '@/lib/types';
-import { defaultPlanningDates, defaultSharedCriteria, isYm, newJob } from '@/lib/types';
+import type { PlanningDates } from '@/lib/types';
+import { defaultPlanningDates } from '@/lib/types';
 import { applyDates, buildTimeline } from '@/lib/timeline';
-import { mergeQueryString, persistState, readInitialState } from '@/lib/urlState';
-import { GOAL_RANGE, MAX_JOBS } from '@/data/pointsCriteria';
+import { mergeQueryString, persistDates, readInitialState } from '@/lib/urlState';
+import { GOAL_RANGE } from '@/data/pointsCriteria';
 import '@/app/i18n/client';
 
 function PageSkeleton() {
@@ -34,83 +29,26 @@ function PageSkeleton() {
     <div className="max-w-[780px] mx-auto px-[26px] pt-[34px]">
       <div className="h-4 w-28 mb-[72px]" style={{ backgroundColor: 'var(--hair)' }} />
       <div className="h-10 w-3/4 mb-[60px]" style={{ backgroundColor: 'var(--hair)' }} />
-      <div className="grid gap-x-9 gap-y-[30px]" style={{ gridTemplateColumns: 'repeat(auto-fill, minmax(min(290px, 100%), 1fr))' }}>
-        {Array.from({ length: 4 }).map((_, i) => (
-          <div key={i}>
-            <div className="h-3 w-24 mb-2.5" style={{ backgroundColor: 'var(--hair)' }} />
-            <div className="h-11" style={{ backgroundColor: 'var(--hair)', opacity: 0.5 }} />
-          </div>
-        ))}
-      </div>
+      <div className="h-40" style={{ backgroundColor: 'var(--hair)', opacity: 0.5 }} />
     </div>
   );
 }
 
-const DEFAULT_JOB_UI: JobUIState = { q: '', open: false };
-
-interface JobCardSlotProps {
-  jobId: string;
-  job: JobAssessment;
-  evaluation: JobEvaluation;
-  canRemove: boolean;
-  ui: JobUIState;
-  openSelect: string | null;
-  setOpenSelect: (key: string | null) => void;
-  ausWorkLocked: boolean;
-  overseasWorkLocked: boolean;
-  collapsed: boolean;
-  patchJob: (id: string, patch: Partial<JobAssessment>) => void;
-  patchJobUI: (id: string, patch: Partial<JobUIState>) => void;
-  removeJob: (id: string) => void;
-  toggleJob: (id: string) => void;
-}
-
-// memo gate for the accordion cards: per-job closures are bound *below* the
-// prop comparison, so a keystroke in one card no longer reconciles the rest
-const JobCardSlot = memo(function JobCardSlot({
-  jobId, job, evaluation, canRemove, ui, openSelect, setOpenSelect,
-  ausWorkLocked, overseasWorkLocked, collapsed,
-  patchJob, patchJobUI, removeJob, toggleJob,
-}: JobCardSlotProps) {
-  return (
-    <JobCard
-      job={job}
-      evaluation={evaluation}
-      canRemove={canRemove}
-      ui={ui}
-      onPatch={(patch) => patchJob(jobId, patch)}
-      onUIPatch={(patch) => patchJobUI(jobId, patch)}
-      onRemove={() => removeJob(jobId)}
-      openSelect={openSelect}
-      setOpenSelect={setOpenSelect}
-      ausWorkLocked={ausWorkLocked}
-      overseasWorkLocked={overseasWorkLocked}
-      collapsed={collapsed}
-      onToggleCollapse={() => toggleJob(jobId)}
-    />
-  );
-});
-
 const PageContent = () => {
-  const { t, ready } = useTranslation();
+  const { ready, t } = useTranslation();
 
   const [mounted, setMounted] = useState(false);
   useEffect(() => setMounted(true), []);
 
+  // shared/jobs are read-only here — they belong to the Profile page.
+  // This page only ever mutates `dates.visa491Grant` (the PR-pathway grant
+  // month) and its own `goalPoints`.
   const initial = useMemo(() => readInitialState(), []);
-  const [shared, setShared] = useState<SharedCriteria>(initial.shared);
-  const [jobs, setJobs] = useState<JobAssessment[]>(initial.jobs);
-  // dates wired to UI in Task 4; initialised here for URL/storage round-trip
+  const shared = initial.shared;
+  const jobs = initial.jobs;
   const [dates, setDates] = useState<PlanningDates>(initial.dates);
   const [goalPoints, setGoalPoints] = useLocalStorage<number>('eoi-goal', GOAL_RANGE.min);
 
-  const [openSelect, setOpenSelect] = useState<string | null>(null);
-  const [jobUI, setJobUI] = useState<Record<string, JobUIState>>({});
-  // Accordion: the one assessment currently expanded for editing.
-  // '__init' resolves to the first job at render time — capturing initial.jobs[0].id
-  // here can mismatch under StrictMode double-invocation (ids are regenerated).
-  const [openJobIdRaw, setOpenJobId] = useState<string | null>('__init');
-  const openJobId = openJobIdRaw === '__init' ? (jobs[0]?.id ?? null) : openJobIdRaw;
   const [copied, setCopied] = useState(false);
   const [exportOpen, setExportOpen] = useState(false);
   const [chipShown, setChipShown] = useState(false);
@@ -122,7 +60,7 @@ const PageContent = () => {
     return `${d.getFullYear()}-${String(d.getMonth() + 1).padStart(2, '0')}`;
   }, []);
 
-  // evaluation now uses date-derived brackets
+  // evaluation uses date-derived brackets
   const derived = useMemo(() => applyDates(shared, jobs, dates, today), [shared, jobs, dates, today]);
   const evaluation = useMemo(() => evaluate(derived.shared, derived.jobs), [derived]);
   const timeline = useMemo(() => buildTimeline({ shared, jobs, dates, today }), [shared, jobs, dates, today]);
@@ -130,47 +68,25 @@ const PageContent = () => {
   const bareScore = evaluation.bareScore;
   const displayTotal = useAnimatedNumber(bareScore);
 
-  // Persist to localStorage + keep the URL shareable (debounced)
+  // Persist the one field this page can edit, and keep the URL shareable
+  // (debounced) — the full state (shared+jobs+dates) still round-trips
+  // through the query string so "copy link" reproduces the whole result.
   const firstRender = useRef(true);
   useEffect(() => {
     if (firstRender.current) { firstRender.current = false; return; }
     const id = setTimeout(() => {
-      persistState(shared, jobs, dates);
+      persistDates(dates);
       const qs = mergeQueryString(window.location.search, shared, jobs, dates);
       window.history.replaceState(null, '', window.location.pathname + (qs ? `?${qs}` : '') + window.location.hash);
     }, 250);
     return () => clearTimeout(id);
   }, [shared, jobs, dates]);
 
-  // Close dropdowns on outside click / Escape
+  // Close the export modal on Escape
   useEffect(() => {
-    const closeAll = () => {
-      setOpenSelect(null);
-      setJobUI((prev) => {
-        if (!Object.values(prev).some((u) => u?.open)) return prev;
-        const next: Record<string, JobUIState> = {};
-        for (const [k, u] of Object.entries(prev)) next[k] = { ...u, open: false };
-        return next;
-      });
-    };
-    const onDocClick = (e: MouseEvent) => {
-      const target = e.target as HTMLElement | null;
-      if (!target?.closest?.('[data-dd]')) closeAll();
-    };
-    const onKey = (e: KeyboardEvent) => {
-      if (e.key === 'Escape') {
-        setExportOpen((open) => {
-          if (!open) closeAll();
-          return false;
-        });
-      }
-    };
-    document.addEventListener('click', onDocClick, true);
+    const onKey = (e: KeyboardEvent) => { if (e.key === 'Escape') setExportOpen(false); };
     document.addEventListener('keydown', onKey);
-    return () => {
-      document.removeEventListener('click', onDocClick, true);
-      document.removeEventListener('keydown', onKey);
-    };
+    return () => document.removeEventListener('keydown', onKey);
   }, []);
 
   // Floating chip appears once the results band scrolls out of view
@@ -185,37 +101,9 @@ const PageContent = () => {
     return () => io.disconnect();
   }, [ready, mounted]);
 
-  const patchShared = useCallback((patch: Partial<SharedCriteria>) => {
-    setShared((prev) => ({ ...prev, ...patch }));
-  }, []);
-
   const patchDates = useCallback((patch: Partial<PlanningDates>) => {
     setDates((prev) => ({ ...prev, ...patch }));
   }, []);
-
-  const patchJob = useCallback((id: string, patch: Partial<JobAssessment>) => {
-    setJobs((prev) => prev.map((j) => (j.id === id ? { ...j, ...patch } : j)));
-  }, []);
-
-  const patchJobUI = useCallback((id: string, patch: Partial<JobUIState>) => {
-    setJobUI((prev) => {
-      const current: JobUIState = prev[id] ?? { q: '', open: false };
-      return { ...prev, [id]: { ...current, ...patch } };
-    });
-  }, []);
-
-  const removeJob = useCallback((id: string) => {
-    setJobs((prev) => prev.filter((j) => j.id !== id));
-    if (openJobId === id) setOpenJobId(null);
-  }, [openJobId]);
-
-  const toggleJob = useCallback((id: string) => {
-    setOpenSelect(null);
-    // Close this card's occupation search too, so the collapsing body
-    // can clip cleanly (overflow flips to hidden mid-transition).
-    patchJobUI(id, { open: false });
-    setOpenJobId(openJobId === id ? null : id);
-  }, [openJobId, patchJobUI]);
 
   const goalDec = useCallback(() => {
     setGoalPoints((prev: number) => Math.max(GOAL_RANGE.min, prev - GOAL_RANGE.step));
@@ -237,11 +125,6 @@ const PageContent = () => {
   }, []);
 
   const handleReset = useCallback(() => {
-    const nj = newJob();
-    setShared({ ...defaultSharedCriteria });
-    setJobs([nj]);
-    setJobUI({});
-    setOpenJobId(nj.id);
     setDates({ ...defaultPlanningDates });
     setGoalPoints(GOAL_RANGE.min);
   }, [setGoalPoints]);
@@ -260,13 +143,12 @@ const PageContent = () => {
 
   if (!ready || !mounted) return <PageSkeleton />;
 
-  const anySearchOpen = Object.values(jobUI).some((u) => u?.open);
-  const chipVisible = chipShown && bareScore > 0 && !exportOpen && !openSelect && !anySearchOpen;
-  const sec2Active = (openSelect !== null && !openSelect.startsWith('sh:')) || anySearchOpen;
+  const chipVisible = chipShown && bareScore > 0 && !exportOpen;
+  const hasProfile = jobs.some((j) => j.anzsco !== '');
 
   return (
     <div
-      className="max-w-[780px] wide:max-w-[1280px] mx-auto"
+      className="max-w-[780px] mx-auto"
       // Landscape on notched phones: keep content clear of the sensor housing
       style={{
         paddingLeft: 'max(26px, env(safe-area-inset-left))',
@@ -275,108 +157,43 @@ const PageContent = () => {
     >
       <Header />
 
-      {/* ≥wide: inputs flow in the left column, the results band sits sticky on the right.
-          Below the breakpoint the grid collapses and DOM order (01·02·03·04) is the layout. */}
-      {/* No items-start here: the right grid item must stretch to the full row
-          height or the sticky panel has no room to travel. */}
-      <div className="wide:grid wide:grid-cols-[minmax(0,1fr)_400px] wide:gap-x-16">
-        <div className="min-w-0">
-          <SharedCriteriaSection
-            shared={derived.shared}
-            onPatch={patchShared}
-            openSelect={openSelect}
-            setOpenSelect={setOpenSelect}
-            ageLocked={isYm(dates.birth)}
-            dates={dates}
-            onDatesPatch={patchDates}
-            today={today}
-          />
-
-          {/* 02 — Skills assessments */}
-          <section
-            className="mt-[72px] relative"
-            style={{ zIndex: sec2Active ? 30 : 'auto', animation: 'eoiFadeUp 0.7s ease 0.16s backwards' }}
-          >
-        <SectionHeading num="02" title={t('sections.jobs')} side="ASSESSMENTS" />
-        <p className="mt-3.5 mb-0 text-[0.78125rem] leading-[1.7] max-w-[46em]" style={{ color: 'var(--muted)' }}>
-          {t('jobsNote')}
+      {!hasProfile && (
+        <p className="mt-[26px] mb-0 text-[0.78125rem] leading-[1.7]" style={{ color: 'var(--muted)' }}>
+          {t('homeNoProfile')}{' '}
+          <Link href="/profile" className="underline underline-offset-4 hover:text-[var(--ink)]" style={{ color: 'var(--ink-soft)', textDecorationColor: 'var(--hair)' }}>
+            {t('navProfile')} →
+          </Link>
         </p>
+      )}
 
-        {jobs.map((job, i) => (
-          <JobCardSlot
-            key={job.id}
-            jobId={job.id}
-            job={derived.jobs[i]}
-            evaluation={evaluation.jobs[i]}
-            canRemove={jobs.length > 1}
-            ui={jobUI[job.id] ?? DEFAULT_JOB_UI}
-            openSelect={openSelect}
-            setOpenSelect={setOpenSelect}
-            ausWorkLocked={isYm(jobs[i].ausWorkStart)}
-            overseasWorkLocked={isYm(jobs[i].overseasWorkStart)}
-            collapsed={openJobId !== job.id}
-            patchJob={patchJob}
-            patchJobUI={patchJobUI}
-            removeJob={removeJob}
-            toggleJob={toggleJob}
-          />
-        ))}
-
-        <ComparisonTable evaluation={evaluation} />
-
-            {jobs.length < MAX_JOBS && (
-              <button
-                type="button"
-                onClick={() => {
-                  const nj = newJob();
-                  setJobs((prev) => [...prev, nj]);
-                  setOpenJobId(nj.id);
-                }}
-                className="w-full mt-[18px] p-[15px] cursor-pointer text-[0.78125rem] tracking-[0.14em] hover:bg-[var(--hover)] hover:border-[var(--ink)]"
-                style={{
-                  background: 'none',
-                  border: '1px dashed var(--muted)',
-                  color: 'var(--ink-soft)',
-                  transition: 'border-color 0.2s ease, background 0.2s ease',
-                }}
-              >
-                +&nbsp;&nbsp;{t('addJob')}
-              </button>
-            )}
-          </section>
-        </div>
-
-        {/* One compact results band everywhere: in document order on narrow
-            screens, a sticky right panel on wide ones. */}
-        <div className="min-w-0 wide:col-start-2 wide:row-start-1 wide:row-span-2">
-          {/* top offset clears the sticky nav bar */}
-          <div className="wide:sticky wide:top-[74px]">
-            <ResultsBand
-              evaluation={evaluation}
-              shared={shared}
-              goal={goalPoints}
-              displayTotal={displayTotal}
-              onGoalDec={goalDec}
-              onGoalInc={goalInc}
-              onOpenExport={openExport}
-              onCopyLink={handleCopyLink}
-              copied={copied}
-              onReset={handleReset}
-              bandRef={bandRef}
-            />
-          </div>
-        </div>
-
-        <div className="min-w-0 wide:col-start-1 wide:row-start-2">
-          <TimelineSection
-            dates={dates}
-            jobs={jobs}
-            timeline={timeline}
-            goal={goalPoints}
-            today={today}
-          />
-        </div>
+      {/* The score card leads — it's the answer this page exists to give.
+          Capped narrower than the page column so it reads as a compact
+          card, not a full-width slab; it never needed the 2-column sticky
+          layout the old single page used when a much taller "shared
+          criteria + jobs" column sat next to it. */}
+      <div className="max-w-[480px] mt-[26px]">
+        <ResultsBand
+          evaluation={evaluation}
+          shared={shared}
+          goal={goalPoints}
+          displayTotal={displayTotal}
+          onGoalDec={goalDec}
+          onGoalInc={goalInc}
+          onOpenExport={openExport}
+          onCopyLink={handleCopyLink}
+          copied={copied}
+          onReset={handleReset}
+          bandRef={bandRef}
+        />
       </div>
+
+      <TimelineSection
+        dates={dates}
+        jobs={jobs}
+        timeline={timeline}
+        goal={goalPoints}
+        today={today}
+      />
 
       <ReferenceSection evaluation={evaluation} />
       <FeeEstimateSection evaluation={evaluation} shared={shared} />
