+++ src/components/EmployerSponsorshipSection.tsx
@@ -25,22 +25,10 @@ interface EmployerSponsorshipSectionProps {
 
 const fmt = (n: number) => `$${n.toLocaleString('en-US')}`;
 
-function ProfileRow({ label, children }: { label: string; children: React.ReactNode }) {
-  return (
-    <div
-      className="grid gap-4 py-[11px] items-baseline"
-      style={{ gridTemplateColumns: '140px 1fr', borderBottom: '1px solid var(--hair-soft)' }}
-    >
-      <span className="text-[0.71875rem] tracking-[0.1em]" style={{ color: 'var(--muted)' }}>{label}</span>
-      <div className="text-[0.8125rem] leading-[1.5]">{children}</div>
-    </div>
-  );
-}
-
 function EditLink() {
   const { t } = useTranslation();
   return (
-    <Link href="/" className="underline underline-offset-4 hover:text-[var(--ink)]" style={{ color: 'var(--muted)', textDecorationColor: 'var(--hair)' }}>
+    <Link href="/profile" className="underline underline-offset-4 hover:text-[var(--ink)]" style={{ color: 'var(--muted)', textDecorationColor: 'var(--hair)' }}>
       {t('spProfileEditLink')}
     </Link>
   );
@@ -109,56 +97,27 @@ function EmployerSponsorshipSection({ jobs, shared, dates, today, inputs, onPatc
 
   return (
     <>
-      {/* 01 — read-only snapshot of what's already entered on the Independent
-          Migration page; this page never edits shared/jobs/dates itself. */}
+      {/* 01 — the sponsorship-specific checklist inputs, local to this page.
+          Age/English/occupations live on the Profile page and are read from
+          there (see the readout line and the per-occupation rows below) —
+          this page never edits shared/jobs/dates itself. */}
       <section className="mt-[58px]" style={{ animation: 'eoiFadeUp 0.7s ease 0.08s backwards' }}>
-        <SectionHeading num="01" title={t('spProfileTitle')} side="PROFILE" />
+        <SectionHeading num="01" title={t('spDetailsTitle')} side="DETAILS" />
         <p className="mt-3.5 mb-0 text-[0.78125rem] leading-[1.7] max-w-[46em]" style={{ color: 'var(--muted)' }}>
           {t('spNote')}
         </p>
 
-        <div className="mt-[22px]">
-          <ProfileRow label={t('compareOcc')}>
-            {jobsWithOcc.length === 0 ? (
-              <span style={{ color: 'var(--muted)' }}>{t('spProfileOccEmpty')} <EditLink /></span>
-            ) : (
-              <div className="flex flex-col gap-1">
-                {jobsWithOcc.map((jr) => {
-                  const occ = findOccupation(jr.job.anzsco);
-                  const tag = String.fromCharCode(65 + jr.index);
-                  return (
-                    <div key={jr.job.id} className="flex items-baseline gap-2.5">
-                      <span className="text-[0.8125rem]" style={{ fontFamily: 'var(--font-serif)' }}>{tag}</span>
-                      <span>{occ ? (lang === 'zh' ? occ.zh : occ.en) : jr.job.anzsco}</span>
-                      <span className="text-xs tabular-nums" style={{ color: 'var(--muted)' }}>{jr.job.anzsco}</span>
-                    </div>
-                  );
-                })}
-              </div>
-            )}
-          </ProfileRow>
-          <ProfileRow label={t('fields.english')}>
-            {evalResult.englishOk ? (
-              t(`options.english.${shared.english}`)
-            ) : (
-              <span style={{ color: 'var(--muted)' }}>{t('spProfileEnglishEmpty')} <EditLink /></span>
-            )}
-          </ProfileRow>
-          <ProfileRow label={t('fields.age')}>
-            {evalResult.ageYears !== null ? (
-              t('spAgeLine', { age: evalResult.ageYears })
-            ) : evalResult.ageUnder45 === true ? (
-              t('spAgeBracketUnder45')
-            ) : (
-              <span style={{ color: 'var(--muted)' }}>{t('spProfileAgeEmpty')} <EditLink /></span>
-            )}
-          </ProfileRow>
-        </div>
-      </section>
-
-      {/* 02 — the sponsorship-specific checklist inputs, local to this page. */}
-      <section className="mt-[58px]" style={{ animation: 'eoiFadeUp 0.7s ease 0.16s backwards' }}>
-        <SectionHeading num="02" title={t('spDetailsTitle')} side="DETAILS" />
+        <p className="mt-2.5 mb-0 text-[0.78125rem] leading-[1.6]" style={{ color: 'var(--ink-soft)' }}>
+          {evalResult.ageYears !== null
+            ? t('spAgeLine', { age: evalResult.ageYears })
+            : evalResult.ageUnder45 === true
+              ? t('spAgeBracketUnder45')
+              : t('spAgeUnknown')}
+          {' · '}
+          {evalResult.englishOk ? t('spEnglishOk') : t('spEnglishUnknown')}
+          {' · '}
+          <EditLink />
+        </p>
 
         {/* One coherent vertical list — a 3-across grid here would fragment
             unevenly on wide screens since the salary picker's natural height
@@ -193,9 +152,9 @@ function EmployerSponsorshipSection({ jobs, shared, dates, today, inputs, onPatc
         </div>
       </section>
 
-      {/* 03 — per-occupation eligibility across all four streams. */}
+      {/* 02 — per-occupation eligibility across all four streams. */}
       <section className="mt-[58px]" style={{ animation: 'eoiFadeUp 0.7s ease 0.24s backwards' }}>
-        <SectionHeading num="03" title={t('spEligibilityTitle')} side="ELIGIBILITY" />
+        <SectionHeading num="02" title={t('spEligibilityTitle')} side="ELIGIBILITY" />
 
         {jobsWithOcc.length === 0 ? (
           <p className="mt-[22px] mb-0 text-[0.78125rem]" style={{ color: 'var(--muted)' }}>{t('spEmpty')} <EditLink /></p>
